requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
idna==3.10
peewee==3.17.7
pyTelegramBotAPI==4.23.0
python-dotenv==0.21.1
python-telegram-bot-pagination==0.0.3
requests==2.32.3
ruff==0.6.9
urllib3==2.2.3
